--- api/v1/cars/urls.py
@@ -1,12 +1,0 @@
-from django.urls import path
-
-from .views import TransportDetail, TransportList  # CarsList,; CarsDetail
-
-urlpatterns = [
-    path('', TransportList.as_view(), name='transport-list'),
-    path('<int:pk>/', TransportDetail.as_view(), name='transport-detail'),
-
-    # Отключены на MVP @OrdinaryWorker
-    # path('', CarsList.as_view(), name='car-list'),
-    # path('<int:pk>/', CarsDetail.as_view(), name='car-detail'),
-]
