+++ api/v1/urls.py
@@ -1,7 +1,6 @@
 from django.urls import include, path
 
 from .autoservice import urls as urls_autoservice
-from .cars import urls as urls_cars
 from .core import urls as urls_core
 # from .users import urls as urls_users
 from .jobs import urls as urls_jobs
@@ -9,7 +8,6 @@ from .jobs import urls as urls_jobs
 urlpatterns = [
     path('autoservice/', include(urls_autoservice)),
     path('core/', include(urls_core)),
-    path('car_models/', include(urls_cars)),
     path('jobs/', include(urls_jobs)),
 
     # Отключены на MVP @OrdinaryWorker
