+++ autoservice/admin.py
@@ -9,6 +9,7 @@ from autoservice.models import (
     GeolocationAutoService,
     GeolocationCity,
     Job,
+    Transport,
     WorkingTime,
     WorkTimeRange,
 )
@@ -58,6 +59,15 @@ class GeolocationCityAdmin(admin.ModelAdmin):
     )
 
 
+class TransportAdmin(admin.ModelAdmin):
+    list_display = ('id',
+                    'brand',
+                    'model'
+                    )
+    list_filter = ('brand', )
+    search_fields = ('brand', )
+
+
 class AutoServiceAdmin(admin.ModelAdmin):
     """
     Кастомизация модели AutoService в админ панеле.
@@ -78,6 +88,7 @@ admin.site.register(Company, CompanyAdmin)
 admin.site.register(City, CityAdmin)
 admin.site.register(Feedback)
 admin.site.register(Job)
+admin.site.register(Transport)
 admin.site.register(WorkTimeRange)
 admin.site.register(WorkingTime)
 admin.site.register(
